// apps/web/src/lib/api/generated/model/promptOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export interface PromptOut {
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  amount: string;
  completed_at: string | null;
  created_at: string;
  id: string;
  invoice_id: string | null;
  payment_id: string | null;
  phone: string;
  receipt: string | null;
  result_desc: string | null;
  /** pending, paid, cancelled, failed or expired */
  status: string;
}